Verify strscan version is a legit version before trying to install it
Fixes #39

This commit patches the `update_global_strscan` task for situations where the command to get the current version of strscan returns a string that is more than a version, e.g., a giant RVM error blob. This situation should not happen often but the extra safeguard is easy to add, especially given how often we rebuild VMs, and should not break current behavior.